🎨 Organize code more
