+++ Loop/Core/Multitouch/MultitouchTrigger+Magnify.swift
@@ -1,0 +1,168 @@
+//
+//  MultitouchTrigger+Magnify.swift
+//  Loop
+//
+//  Created by Kai Azim on 2026-01-30.
+//
+
+import CoreGraphics
+import Subsurface
+
+/// Magnify handling lives separately from the main trigger as it has its own
+/// activation threshold and reversal model, including direction changes between
+/// magnify-in and magnify-out gestures and radial-menu center commits
+extension MultitouchTrigger {
+    func handleMagnify(_ magnify: SubsurfaceGestureEvent.MagnifyEvent, fingerCount: Int) async {
+        guard let entry = recognizerRegistry.entry(for: fingerCount) else { return }
+
+        // Radial menu magnify triggers the center action regardless of direction.
+        if let radialMenuGesture = entry.radialMenuGesture {
+            await handleRadialMenuMagnify(magnify, fingerCount: fingerCount, gesture: radialMenuGesture)
+            return
+        }
+
+        switch magnify.phase {
+        case .began, .changed:
+            if magnify.phase == .began,
+               recognizerRegistry.session(for: fingerCount)?.hasGestureBegun != true {
+                recognizerRegistry.session(for: fingerCount)?.reset()
+            }
+            guard let session = recognizerRegistry.session(for: fingerCount), !session.isGestureRejected else { return }
+
+            if !session.hasGestureBegun {
+                let initialGesture = magnify.distance >= magnify.originDistance ? entry.magnifyOutGesture : entry.magnifyInGesture
+                guard let initialGesture else {
+                    return
+                }
+                handleGestureBegan(fingerCount: fingerCount, gesture: initialGesture)
+                recognizerRegistry.session(for: fingerCount)?.setResolvedGesture(initialGesture)
+            }
+
+            guard let session = recognizerRegistry.session(for: fingerCount), !session.isGestureRejected,
+                  let activeGesture = session.resolvedGesture
+            else {
+                return
+            }
+
+            guard await activateGestureIfNeeded(
+                fingerCount: fingerCount,
+                magnifyDisplacement: magnify.distance - magnify.originDistance,
+                magnifyActivationThreshold: magnifyCycleStepSize
+            ),
+                let session = recognizerRegistry.session(for: fingerCount),
+                !session.isGestureRejected
+            else {
+                return
+            }
+
+            if magnifyReversalDetected(currentGesture: activeGesture, magnify: magnify) {
+                let opposite = activeGesture.kind == .magnifyIn ? entry.magnifyOutGesture : entry.magnifyInGesture
+                handleMagnifyReversal(
+                    fingerCount: fingerCount,
+                    currentGesture: activeGesture,
+                    oppositeGesture: opposite,
+                    distance: magnify.distance
+                )
+                return
+            }
+
+            let canRepeatAction = resolvedWindowAction(from: activeGesture).map {
+                $0.canRepeat || $0.direction == .cycle
+            } ?? false
+
+            session.commitMagnify(
+                gesture: activeGesture,
+                distance: magnify.distance,
+                step: magnifyCycleStepSize,
+                canRepeat: canRepeatAction
+            ) { reverse in
+                triggerSingleAction(from: activeGesture, reverse: reverse)
+            }
+
+            if let window = session.pendingTargetWindow,
+               resolvedWindowAction(from: activeGesture)?.canRepeat == true {
+                targetResolver.rememberRepeatableWindow(window, canRepeat: true)
+            }
+
+        case .ended, .cancelled:
+            resetLoopState(for: fingerCount)
+
+        default:
+            break
+        }
+    }
+
+    /// Magnify within a radial menu gesture triggers the center (last) radial menu action.
+    private func handleRadialMenuMagnify(
+        _ magnify: SubsurfaceGestureEvent.MagnifyEvent,
+        fingerCount: Int,
+        gesture: GestureBinding
+    ) async {
+        switch magnify.phase {
+        case .began, .changed:
+            if magnify.phase == .began, recognizerRegistry.session(for: fingerCount)?.hasGestureBegun != true {
+                handleGestureBegan(fingerCount: fingerCount, gesture: gesture)
+            }
+            guard await activateGestureIfNeeded(
+                fingerCount: fingerCount,
+                magnifyDisplacement: magnify.distance - magnify.originDistance
+            ) else { return }
+            guard let session = recognizerRegistry.session(for: fingerCount), !session.isGestureRejected else { return }
+
+            let actions = radialMenuActions
+            guard !actions.isEmpty else { return }
+            let centerActionIndex = actions.count - 1
+
+            session.commitRadialMagnify(
+                distance: magnify.distance,
+                originDistance: magnify.originDistance,
+                step: magnifyCycleStepSize
+            ) { reverse in
+                triggerRadialMenuAction(at: centerActionIndex, from: actions[...], reverse: reverse)
+            }
+
+        case .ended, .cancelled:
+            resetLoopState(for: fingerCount)
+
+        default:
+            break
+        }
+    }
+
+    private func magnifyReversalDetected(
+        currentGesture: GestureBinding,
+        magnify: SubsurfaceGestureEvent.MagnifyEvent
+    ) -> Bool {
+        switch currentGesture.kind {
+        case .magnifyIn:
+            magnify.distance >= magnify.originDistance
+        case .magnifyOut:
+            magnify.distance <= magnify.originDistance
+        default:
+            false
+        }
+    }
+
+    private func handleMagnifyReversal(
+        fingerCount: Int,
+        currentGesture: GestureBinding,
+        oppositeGesture: GestureBinding?,
+        distance: CGFloat
+    ) {
+        if let oppositeGesture {
+            guard let session = recognizerRegistry.session(for: fingerCount) else { return }
+            session.switchMagnifyGesture(to: oppositeGesture, distance: distance)
+            triggerSingleAction(from: oppositeGesture, reverse: false)
+
+            if let window = session.pendingTargetWindow,
+               resolvedWindowAction(from: oppositeGesture)?.canRepeat == true {
+                targetResolver.rememberRepeatableWindow(window, canRepeat: true)
+            }
+        } else if isCycleAction(currentGesture) {
+            triggerSingleAction(from: currentGesture, reverse: true)
+            recognizerRegistry.session(for: fingerCount)?.updateLastCommitMagnifyDistance(distance)
+        } else {
+            resetLoopState(for: fingerCount, forceClose: true)
+        }
+    }
+}
